Some refactoring and error handling

+++ src/AppBundle/Controller/FriendController.php
@@ -2,13 +2,10 @@
 
 namespace AppBundle\Controller;
 
-use Doctrine\Common\Collections\Criteria;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use AppBundle\Entity\AppUser;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Component\HttpFoundation\Request;
-use AppBundle\Entity\Friendship;
-use AppBundle\Entity\FriendshipType;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use AppBundle\Entity\AppUser;
 
 class FriendController extends Controller
 {
@@ -21,23 +18,115 @@ class FriendController extends Controller
         {
             $appUser = $this->getDoctrine()->getRepository('AppBundle:AppUser')->find($slug);
 
-            $friendships = $appUser->getFriendships();
-
-            $friendArray = array();
-
-            foreach($friendships as $friendship)
+            if($appUser === null)
             {
-                $fType = $friendship->getFriendshipType()->getFshipType();
-                if ($fType == "Accepted")
+                throw $this->createNotFoundException();
+            }
+            else
+            {
+                $friendships = $appUser->getFriendships();
+
+                $friendArray = array();
+
+                foreach ($friendships as $friendship)
                 {
-                    array_push($friendArray, $friendship->getFriendUser());
+                    $fType = $friendship->getFriendshipType()->getFshipType();
+                    if ($fType == "Accepted")
+                    {
+                        array_push($friendArray, $friendship->getFriendUser());
+                    }
+                }
+
+                return $this->render(
+                    'Friend/friendList.html.twig',
+                    array('friends' => $friendArray)
+                );
+            }
+        }
+        else
+        {
+            throw $this->createAccessDeniedException();
+        }
+    }
+
+    /**
+     * @Route("users/{slug}", name="user_show")
+     */
+    public function showUserAction($slug)
+    {
+        if($this->get('security.authorization_checker')->isGranted('IS_AUTHENTICATED_FULLY'))
+        {
+            $appUser = $this->getDoctrine()->getRepository('AppBundle:AppUser')->find($slug);
+
+            if($appUser === null)
+            {
+                throw $this->createNotFoundException();
+            }
+            else
+            {
+                $loggedInUser = $this->getDoctrine()->getRepository('AppBundle:AppUser')->find($this->getUser()->getId());
+
+                if ($appUser == $loggedInUser)
+                {
+                    return $this->redirectToRoute('profile_show');
+                }
+                else
+                {
+                    if ($loggedInUser->getFriendships()->isEmpty())
+                    {
+                        return $this->render(
+                            'Friend/user.html.twig',
+                            array('app_user' => $appUser)
+                        );
+                    }
+                    else
+                    {
+                        foreach ($loggedInUser->getFriendships() as $friendship)
+                        {
+                            if ($friendship->getFriendUser() == $appUser)
+                            {
+                                if ($friendship->getFriendshipType()->getFshipType() == "Pending")
+                                {
+                                    return $this->render(
+                                        'Friend/pendingUser.html.twig',
+                                        array('app_user' => $appUser)
+                                    );
+                                }
+                                elseif ($friendship->getFriendshipType()->getFshipType() == "Accepted")
+                                {
+                                    $friendUpdates = $appUser->getStatusUpdates()->toArray();
+
+                                    usort($friendUpdates, function ($a, $b)
+                                    {
+                                        return $b->getCreationDate()->format('U') - $a->getCreationDate()->format('U');
+                                    });
+
+                                    return $this->render(
+                                        'Friend/friend.html.twig',
+                                        array('app_user' => $appUser, 'updates' => $friendUpdates)
+                                    );
+                                }
+                                elseif ($friendship->getFriendshipType()->getFshipType() == "Asked")
+                                {
+                                    return $this->render(
+                                        'Friend/askedUser.html.twig',
+                                        array('app_user' => $appUser)
+                                    );
+                                }
+                            }
+                            else
+                            {
+                                continue;
+                            }
+                        }
+
+                        return $this->render(
+                            'Friend/user.html.twig',
+                            array('app_user' => $appUser)
+                        );
+                    }
                 }
             }
-
-            return $this->render(
-                'Friend/friendList.html.twig', array(
-                'friends' => $friendArray,
-            ));
         }
         else
         {
@@ -56,36 +145,42 @@ class FriendController extends Controller
 
             if ($post->request->has('submit'))
             {
-                $appUserID = $this->getUser()->getId();
-
-                $appUser = $this->getDoctrine()->getRepository('AppBundle:AppUser')->find($appUserID);
-
                 $friendUser = $this->getDoctrine()->getRepository('AppBundle:AppUser')->find($slug);
 
-                $friendships = $appUser->getFriendships();
-
-                $em = $this->getDoctrine()->getManager();
-
-                foreach($friendships as $friendship)
+                if($friendUser === null)
                 {
-                    if($friendship->getFriendUser() == $friendUser)
-                    {
-                        $em->remove($friendship);
-                    }
+                    throw $this->createNotFoundException();
                 }
-
-                $friendships2 = $friendUser->getFriendships();
-
-                foreach($friendships2 as $friendship2)
+                else
                 {
-                    if($friendship2->getFriendUser() == $appUser)
-                    {
-                        $em->remove($friendship2);
-                    }
-                }
+                    $appUser = $this->getDoctrine()->getRepository('AppBundle:AppUser')->find($this->getUser()->getId());
 
-                $em->flush();
-                return $this->redirectToRoute('users');
+                    $friendships = $appUser->getFriendships();
+
+                    $em = $this->getDoctrine()->getManager();
+
+                    foreach ($friendships as $friendship)
+                    {
+                        if ($friendship->getFriendUser() == $friendUser)
+                        {
+                            $em->remove($friendship);
+                        }
+                    }
+
+                    $friendships2 = $friendUser->getFriendships();
+
+                    foreach ($friendships2 as $friendship2)
+                    {
+                        if ($friendship2->getFriendUser() == $appUser)
+                        {
+                            $em->remove($friendship2);
+                        }
+                    }
+
+                    $em->flush();
+
+                    return $this->redirect($this->generateUrl('user_show', array('slug' => $slug)));
+                }
             }
         }
         else
